+++ packages/wallet-engine/src/engine/approve.ts
@@ -18,8 +18,6 @@ type ProviderRequest<T = unknown, U = unknown> = WithPreopenedInstanceId & {
   proceed: () => Promise<Pick<PendingJsonRpcResponse<U>, 'error' | 'result'>>;
 };
 
-export type PersonalSignRequest = ProviderRequest<[string, string, KeyType], string>;
-
 export type IncomingTransaction = {
   from: string;
   to: string;
@@ -31,7 +29,10 @@ export type IncomingTransaction = {
   maxPriorityFeePerGas?: string;
 };
 
+export type PayloadSignData = Record<string, any>;
 export type SendTransactionRequest = ProviderRequest<[IncomingTransaction], string>;
+export type PersonalSignRequest = ProviderRequest<[string, string, KeyType], string>;
+export type PayloadSignRequest<T = PayloadSignData> = ProviderRequest<[T, KeyType], string>;
 
 export type TransferParams = {
   from: string;
@@ -46,6 +47,7 @@ export type ApproveEngineOptions = {
   onSendTransaction?: (request: SendTransactionRequest) => Promise<void>;
   onTransfer?: (request: TransferRequest) => Promise<void>;
   onPersonalSign?: (request: PersonalSignRequest) => Promise<void>;
+  onPayloadSign?: (request: PayloadSignRequest) => Promise<void>;
 };
 
 type RequestMiddleware<T, U> = (
@@ -73,6 +75,7 @@ const createRequestMiddleware = <T = any, U = any>(handler: RequestMiddleware<T,
 
 const noop = async () => {};
 export const createApproveEngine = ({
+  onPayloadSign = noop,
   onPersonalSign = noop,
   onSendTransaction = noop,
   onTransfer = noop,
@@ -89,12 +92,37 @@ export const createApproveEngine = ({
         });
       }),
 
+      /**
+       * @deprecated Use `ed25519_signRaw` instead. This method is unsafe and should not be used.
+       *
+       * TODO: Remove this method after migrating `@cere/embed-wallet-inject` to use `ed25519_signRaw` and `ed25519_signPayload`.
+       */
       ed25519_sign: createRequestMiddleware<[string, string]>(async (req, proceed) => {
         const [account, message] = req.params!;
 
         await onPersonalSign({
           preopenInstanceId: req.preopenInstanceId,
           params: [message, account, 'ed25519' as KeyType],
+          proceed,
+        });
+      }),
+
+      ed25519_signRaw: createRequestMiddleware<[string, string]>(async (req, proceed) => {
+        const [account, message] = req.params!;
+
+        await onPersonalSign({
+          preopenInstanceId: req.preopenInstanceId,
+          params: [message, account, 'ed25519' as KeyType],
+          proceed,
+        });
+      }),
+
+      ed25519_signPayload: createRequestMiddleware<[PayloadSignData]>(async (req, proceed) => {
+        const [payload] = req.params!;
+
+        await onPayloadSign({
+          preopenInstanceId: req.preopenInstanceId,
+          params: [payload, 'ed25519' as KeyType],
           proceed,
         });
       }),