+++ src/routes/AuthorizationRouter.tsx
@@ -1,7 +1,15 @@
 import { Route, Routes } from 'react-router-dom';
 
-import { Authorize, AuthorizeClose, AuthorizeRedirect, IntroRoute, LoginRoute, OtpRoute } from './Authorize';
 import { useAppContextStore } from '~/hooks';
+import {
+  Authorize,
+  AuthorizeClose,
+  AuthorizeRedirect,
+  AuthorizeIntro,
+  AuthorizeLogin,
+  AuthorizeOtp,
+  AuthorizePermissions,
+} from './Authorize';
 
 export const AuthorizationRouter = () => {
   const store = useAppContextStore();
@@ -11,13 +19,14 @@ export const AuthorizationRouter = () => {
   return (
     <Routes>
       <Route element={<Authorize />}>
-        <Route index element={isGame || skipLoginIntro ? <LoginRoute /> : <IntroRoute />} />
+        <Route index element={isGame || skipLoginIntro ? <AuthorizeLogin /> : <AuthorizeIntro />} />
         <Route path="close" element={<AuthorizeClose />} />
         <Route path="redirect" element={<AuthorizeRedirect />} />
-        <Route path="intro" element={<IntroRoute />} />
-        <Route path="signin" element={<LoginRoute variant="signin" />} />
-        <Route path="signup" element={<LoginRoute variant="signup" />} />
-        <Route path="otp" element={<OtpRoute />} />
+        <Route path="intro" element={<AuthorizeIntro />} />
+        <Route path="signin" element={<AuthorizeLogin variant="signin" />} />
+        <Route path="signup" element={<AuthorizeLogin variant="signup" />} />
+        <Route path="otp" element={<AuthorizeOtp />} />
+        <Route path="permissions" element={<AuthorizePermissions />} />
       </Route>
     </Routes>
   );