+++ src/routes/Authorize/index.ts
@@ -1,6 +1,7 @@
 export { default as Authorize } from './Authorize';
 export { default as AuthorizeClose } from './AuthorizeClose';
 export { default as AuthorizeRedirect } from './AuthorizeRedirect';
-export { IntroRoute } from './IntroRoute';
-export { LoginRoute } from './LoginRoute';
-export { OtpRoute } from './OtpRoute';
+export { default as AuthorizePermissions } from './AuthorizePermissions';
+export { default as AuthorizeIntro } from './AuthorizeIntro';
+export { default as AuthorizeLogin } from './AuthorizeLogin';
+export { default as AuthorizeOtp } from './AuthorizeOtp';