+++ CHANGELOG.md
@@ -2,7 +2,14 @@
 
 ### vNext
 
--
+- 
+
+### v1.30.0
+
+- Add Wallet Permission System ([EIP-2255](https://eips.ethereum.org/EIPS/eip-2255)) support
+- Added `initializing` wallet instance status
+- Add `ed25519_signPayload` and `ed25519_signRaw` RPC methods
+- Updated `@cere/embed-wallet-inject` to use new RPC methods
 
 ### v1.29.4